test: migrate Phase 3 assembly/respondents/targets e2e to component
Move behavioural coverage for backoffice assembly, assembly gsheet CRUD,
assembly user management, backoffice respondents, respondent field schema
and targets pages to tests/component (FakeUnitOfWork). Keep PG smokes in
e2e; mark search/SelectionSettings/attribute-aggregation/sort-order tests
db_semantics. The CSV-upload diff stash is Redis-backed so those diff
tests stay e2e (requires_redis) - doc corrected.
