+++ backend/tests/e2e/test_assembly_gsheet_crud.py
@@ -1,5 +1,7 @@
-"""ABOUTME: End-to-end AssemblyGSheet CRUD tests
-ABOUTME: Tests complete AssemblyGSheet creation, viewing, editing, and deletion workflows through web interface"""
+"""ABOUTME: End-to-end AssemblyGSheet CRUD smoke + DB-semantics tests
+ABOUTME: Keeps create/delete/workflow smokes and SelectionSettings JSON round-trip checks on real PostgreSQL"""
+
+import pytest
 
 from opendlp.domain.selection_settings import DEFAULT_ADDRESS_COLS, DEFAULT_COLS_TO_KEEP, DEFAULT_ID_COLUMN
 from opendlp.service_layer.unit_of_work import SqlAlchemyUnitOfWork
@@ -8,46 +10,6 @@ from tests.e2e.helpers import get_csrf_token
 
 class TestAssemblyGSheetCreateView:
     """Test AssemblyGSheet creation functionality."""
-
-    def test_create_gsheet_get_form_no_existing(self, logged_in_admin, existing_assembly):
-        """Test create gsheet form is displayed when no gsheet exists."""
-        response = logged_in_admin.get(f"/assemblies/{existing_assembly.id}/gsheet")
-        assert response.status_code == 200
-        assert b"Configure Google Spreadsheet" in response.data
-        assert b"Google Spreadsheet URL" in response.data
-        assert b"Team Configuration" in response.data
-        assert b"Save Configuration" in response.data
-
-    def test_create_gsheet_form_contains_all_fields(self, logged_in_admin, existing_assembly):
-        """Test that gsheet form contains all required fields."""
-        response = logged_in_admin.get(f"/assemblies/{existing_assembly.id}/gsheet")
-        assert response.status_code == 200
-
-        # URL field
-        assert b"Google Spreadsheet URL" in response.data
-        assert b'placeholder="https://docs.google.com/spreadsheets/d/' in response.data  # Placeholder
-
-        # Team selection
-        assert b"Team Configuration" in response.data
-        assert b"UK Team" in response.data
-        assert b"EU Team" in response.data
-        assert b"Australia Team" in response.data
-
-        # Tab names
-        assert b"Respondents Tab Name" in response.data
-        assert b"Targets Tab Name" in response.data
-        assert b"Already Selected Tab Name" in response.data
-
-        # ID column
-        assert b"ID Column" in response.data
-
-        # String list fields
-        assert b"Address Columns" in response.data
-        assert b"Columns to Keep" in response.data
-
-        # Checkboxes
-        assert b"Check Same Address" in response.data
-        assert b"Generate Remaining Tab" in response.data
 
     def test_create_gsheet_success(self, logged_in_admin, existing_assembly):
         """Test successful gsheet creation."""
@@ -77,96 +39,11 @@ class TestAssemblyGSheetCreateView:
             flash_messages = [msg[1] for msg in session.get("_flashes", [])]
             assert any("configuration created successfully" in msg for msg in flash_messages)
 
-    def test_create_gsheet_validation_invalid_url(self, logged_in_admin, existing_assembly):
-        """Test form validation errors for invalid URL."""
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet",
-            data={
-                "url": "https://invalid-url.com/not-google-sheets",
-                "team": "uk",
-                "select_registrants_tab": "Respondents",
-                "select_targets_tab": "Categories",
-                "id_column": "nationbuilder_id",
-                "check_same_address_cols_string": "address1, postal_code",
-                "columns_to_keep_string": "name, email",
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet"),
-            },
-        )
-
-        # Should return form with validation errors
-        assert response.status_code == 200
-        assert b"error" in response.data or b"Invalid" in response.data
-
-    def test_create_gsheet_validation_required_fields(self, logged_in_admin, existing_assembly):
-        """Test form validation errors for missing required fields."""
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet",
-            data={
-                "url": "",  # Missing required URL
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet"),
-            },
-        )
-
-        # Should return form with validation errors
-        assert response.status_code == 200
-        assert b"error" in response.data or b"required" in response.data
-
-    def test_create_gsheet_with_custom_team_and_string_fields(self, logged_in_admin, existing_assembly):
-        """Test creating gsheet with 'other' team and custom string field values."""
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/custom123456789/edit",
-                "team": "other",  # Custom configuration
-                "select_registrants_tab": "CustomRespondents",
-                "select_targets_tab": "CustomCategories",
-                "id_column": "custom_id",
-                "check_same_address_cols_string": "street_address, postal_code, city",
-                "columns_to_keep_string": "first_name, last_name, email, phone, address",
-                "check_same_address": True,
-                "generate_remaining_tab": False,
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet"),
-            },
-            follow_redirects=False,
-        )
-
-        # Should succeed
-        assert response.status_code == 302
-        assert f"/assemblies/{existing_assembly.id}/data" in response.location
-
-        # Check flash message
-        with logged_in_admin.session_transaction() as session:
-            flash_messages = [msg[1] for msg in session.get("_flashes", [])]
-            assert any("configuration created successfully" in msg for msg in flash_messages)
-
-    def test_create_gsheet_permission_denied_for_user(self, logged_in_user, existing_assembly):
-        """Test regular users cannot create gsheet configurations."""
-        response = logged_in_user.get(f"/assemblies/{existing_assembly.id}/gsheet")
-        # Should redirect or show permission error
-        assert response.status_code in [302, 403, 500]
-
-    def test_create_gsheet_redirects_when_not_logged_in(self, client, existing_assembly):
-        """Test create gsheet redirects to login when not authenticated."""
-        response = client.get(f"/assemblies/{existing_assembly.id}/gsheet")
-        assert response.status_code == 302
-        assert "login" in response.location
-
 
 class TestAssemblyGSheetEditView:
-    """Test AssemblyGSheet editing functionality."""
+    """Test AssemblyGSheet editing DB semantics on real PostgreSQL."""
 
-    def test_edit_gsheet_get_form_with_existing(self, logged_in_admin, assembly_with_gsheet):
-        """Test edit gsheet form is displayed with existing data."""
-        assembly, gsheet = assembly_with_gsheet
-        response = logged_in_admin.get(f"/assemblies/{assembly.id}/gsheet")
-        assert response.status_code == 200
-        assert b"Edit Google Spreadsheet Configuration" in response.data
-        assert gsheet.url.encode() in response.data
-        assert gsheet.select_registrants_tab.encode() in response.data
-        assert gsheet.select_targets_tab.encode() in response.data
-        assert b"Save Changes" in response.data
-        assert b"Remove Configuration" in response.data
-
+    @pytest.mark.db_semantics
     def test_edit_gsheet_get_form_populates_selection_settings_fields(
         self, logged_in_admin, assembly_with_gsheet, postgres_session_factory
     ):
@@ -203,6 +80,7 @@ class TestAssemblyGSheetEditView:
             "Columns to keep string from SelectionSettings should pre-populate the form"
         )
 
+    @pytest.mark.db_semantics
     def test_edit_gsheet_success_with_team_eu(self, logged_in_admin, assembly_with_gsheet, postgres_session_factory):
         """Test successful gsheet editing with eu team overriding some settings."""
         assembly, _gsheet = assembly_with_gsheet
@@ -250,6 +128,7 @@ class TestAssemblyGSheetEditView:
             assert saved_assembly.selection_settings.check_same_address_cols == DEFAULT_ADDRESS_COLS["eu"]
             assert saved_assembly.selection_settings.columns_to_keep == DEFAULT_COLS_TO_KEEP["eu"]
 
+    @pytest.mark.db_semantics
     def test_edit_gsheet_success_with_team_custom(
         self, logged_in_admin, assembly_with_gsheet, postgres_session_factory
     ):
@@ -301,28 +180,6 @@ class TestAssemblyGSheetEditView:
                 "city",
             ]
 
-    def test_edit_gsheet_validation_errors(self, logged_in_admin, assembly_with_gsheet):
-        """Test form validation errors on gsheet editing."""
-        assembly, _gsheet = assembly_with_gsheet
-        response = logged_in_admin.post(
-            f"/assemblies/{assembly.id}/gsheet",
-            data={
-                "url": "",  # Empty URL should fail validation
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{assembly.id}/gsheet"),
-            },
-        )
-
-        # Should return form with validation errors
-        assert response.status_code == 200
-        assert b"error" in response.data or b"required" in response.data
-
-    def test_edit_gsheet_permission_denied_for_user(self, logged_in_user, assembly_with_gsheet):
-        """Test regular users cannot edit gsheet configurations."""
-        assembly, _gsheet = assembly_with_gsheet
-        response = logged_in_user.get(f"/assemblies/{assembly.id}/gsheet")
-        # Should redirect or show permission error
-        assert response.status_code in [302, 403, 500]
-
 
 class TestAssemblyGSheetDeleteView:
     """Test AssemblyGSheet deletion functionality."""
@@ -346,44 +203,6 @@ class TestAssemblyGSheetDeleteView:
         with logged_in_admin.session_transaction() as session:
             flash_messages = [msg[1] for msg in session.get("_flashes", [])]
             assert any("configuration removed successfully" in msg for msg in flash_messages)
-
-    def test_delete_nonexistent_gsheet(self, logged_in_admin, existing_assembly):
-        """Test deleting non-existent gsheet shows error."""
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet/delete",
-            data={
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet/delete"),
-            },
-            follow_redirects=False,
-        )
-
-        # Should redirect with error message
-        assert response.status_code == 302
-        assert f"/assemblies/{existing_assembly.id}" in response.location
-
-        # Check error flash message
-        with logged_in_admin.session_transaction() as session:
-            flash_messages = [msg[1] for msg in session.get("_flashes", [])]
-            assert any("not found" in msg for msg in flash_messages)
-
-    def test_delete_gsheet_permission_denied_for_user(self, logged_in_user, assembly_with_gsheet):
-        """Test regular users cannot delete gsheet configurations."""
-        assembly, _gsheet = assembly_with_gsheet
-        response = logged_in_user.post(
-            f"/assemblies/{assembly.id}/gsheet/delete",
-            data={
-                "csrf_token": get_csrf_token(logged_in_user, f"/assemblies/{assembly.id}/gsheet/delete"),
-            },
-        )
-        # Should redirect or show permission error
-        assert response.status_code in [302, 403, 500]
-
-    def test_delete_gsheet_redirects_when_not_logged_in(self, client, assembly_with_gsheet):
-        """Test delete gsheet redirects to login when not authenticated."""
-        assembly, _gsheet = assembly_with_gsheet
-        response = client.post(f"/assemblies/{assembly.id}/gsheet/delete")
-        assert response.status_code == 302
-        assert "login" in response.location
 
 
 class TestAssemblyGSheetWorkflowIntegration:
@@ -450,350 +269,3 @@ class TestAssemblyGSheetWorkflowIntegration:
         assert verify_response.status_code == 200
         assert b"Configure Google Spreadsheet" in verify_response.data  # Create form
         assert b"Edit Google Spreadsheet Configuration" not in verify_response.data
-
-    def test_gsheet_state_transitions(self, logged_in_admin, existing_assembly):
-        """Test state transitions between no gsheet -> has gsheet -> no gsheet."""
-        assembly = existing_assembly
-
-        # Initially no gsheet - should show create form
-        initial_response = logged_in_admin.get(f"/assemblies/{assembly.id}/gsheet")
-        assert initial_response.status_code == 200
-        assert b"Configure Google Spreadsheet" in initial_response.data
-        assert b"Save Configuration" in initial_response.data
-
-        # Create gsheet
-        logged_in_admin.post(
-            f"/assemblies/{assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/state123456789/edit",
-                "team": "uk",
-                "select_registrants_tab": "StateRespondents",
-                "select_targets_tab": "StateCategories",
-                "id_column": "state_id",
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{assembly.id}/gsheet"),
-            },
-        )
-
-        # Now has gsheet - should show edit form
-        edit_response = logged_in_admin.get(f"/assemblies/{assembly.id}/gsheet")
-        assert edit_response.status_code == 200
-        assert b"Edit Google Spreadsheet Configuration" in edit_response.data
-        assert b"Save Changes" in edit_response.data
-        assert b"Remove Configuration" in edit_response.data
-        assert b"state123456789" in edit_response.data
-
-        # Delete gsheet
-        logged_in_admin.post(
-            f"/assemblies/{assembly.id}/gsheet/delete",
-            data={
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{assembly.id}/gsheet/delete"),
-            },
-        )
-
-        # Back to no gsheet - should show create form again
-        final_response = logged_in_admin.get(f"/assemblies/{assembly.id}/gsheet")
-        assert final_response.status_code == 200
-        assert b"Configure Google Spreadsheet" in final_response.data
-        assert b"Save Configuration" in final_response.data
-        assert b"Edit Google Spreadsheet Configuration" not in final_response.data
-
-    def test_navigation_breadcrumbs_work(self, logged_in_admin, existing_assembly):
-        """Test that navigation breadcrumbs are functional."""
-        assembly = existing_assembly
-
-        # Create form should have breadcrumbs
-        create_response = logged_in_admin.get(f"/assemblies/{assembly.id}/gsheet")
-        assert create_response.status_code == 200
-        assert b"Dashboard" in create_response.data
-        assert assembly.title.encode() in create_response.data
-
-        # Create gsheet and test edit form breadcrumbs
-        logged_in_admin.post(
-            f"/assemblies/{assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/breadcrumb123456789/edit",
-                "team": "uk",
-                "select_registrants_tab": "BreadcrumbRespondents",
-                "select_targets_tab": "BreadcrumbCategories",
-                "id_column": "breadcrumb_id",
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{assembly.id}/gsheet"),
-            },
-        )
-
-        # Edit form should also have breadcrumbs
-        edit_response = logged_in_admin.get(f"/assemblies/{assembly.id}/gsheet")
-        assert edit_response.status_code == 200
-        assert b"Dashboard" in edit_response.data
-        assert assembly.title.encode() in edit_response.data
-
-
-class TestAssemblyGSheetPermissions:
-    """Test gsheet permission handling."""
-
-    def test_permissions_properly_enforced_for_regular_users(self, logged_in_user, existing_assembly):
-        """Test that permission restrictions are properly enforced for regular users."""
-        assembly = existing_assembly
-
-        # GET request should be blocked
-        get_response = logged_in_user.get(f"/assemblies/{assembly.id}/gsheet")
-        assert get_response.status_code in [302, 403, 500]
-
-        # POST request should also be blocked
-        post_response = logged_in_user.post(
-            f"/assemblies/{assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/unauthorized123456789/edit",
-                "csrf_token": get_csrf_token(logged_in_user, f"/assemblies/{assembly.id}/gsheet"),
-            },
-        )
-        assert post_response.status_code in [302, 403, 500]
-
-    def test_nonexistent_assembly_returns_error(self, logged_in_admin):
-        """Test accessing gsheet for non-existent assembly returns error."""
-        response = logged_in_admin.get("/assemblies/00000000-0000-0000-0000-000000000000/gsheet")
-        assert response.status_code == 302  # Should redirect to dashboard with error
-
-    def test_url_validation_enforced(self, logged_in_admin, existing_assembly):
-        """Test that URL validation is properly enforced."""
-        assembly = existing_assembly
-
-        invalid_urls = [
-            "not-a-url",
-            "https://example.com/not-google-sheets",
-            "https://drive.google.com/file/d/wrong-format",
-            "",
-        ]
-
-        for invalid_url in invalid_urls:
-            response = logged_in_admin.post(
-                f"/assemblies/{assembly.id}/gsheet",
-                data={
-                    "url": invalid_url,
-                    "team": "uk",
-                    "select_registrants_tab": "Test",
-                    "select_targets_tab": "Test",
-                    "id_column": "test_id",
-                    "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{assembly.id}/gsheet"),
-                },
-            )
-            # Should return form with validation error
-            assert response.status_code == 200
-            assert b"error" in response.data or b"Invalid" in response.data or b"required" in response.data
-
-
-class TestAssemblyGSheetValidation:
-    """Test gsheet-specific validation rules."""
-
-    def test_create_gsheet_hard_validation_check_address_without_columns(self, logged_in_admin, existing_assembly):
-        """Test hard validation: check_same_address=True requires address columns."""
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/validation123456789/edit",
-                "team": "other",  # Custom config to avoid team defaults
-                "select_registrants_tab": "Respondents",
-                "select_targets_tab": "Categories",
-                "replace_registrants_tab": "Remaining",
-                "replace_targets_tab": "Replacement Categories",
-                "already_selected_tab": "Selected",
-                "id_column": "test_id",
-                "check_same_address": "y",  # Enabled
-                "check_same_address_cols_string": "",  # Empty - should fail validation
-                "columns_to_keep_string": "first_name, last_name",
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet"),
-            },
-        )
-
-        # Should return form with validation error (not redirect)
-        assert response.status_code == 200
-        # Check for validation error message
-        assert b"address columns" in response.data.lower() or b"must specify" in response.data.lower()
-
-    def test_create_gsheet_hard_validation_passes_when_team_not_other(self, logged_in_admin, existing_assembly):
-        """
-        Test hard validation:
-        - check_same_address=True requires address columns.
-        - team="uk" will auto generate address columns after the form validation.
-        - so overall we should succeed.
-        """
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/validation123456789/edit",
-                "team": "uk",  # uk config to create team defaults
-                "select_registrants_tab": "Respondents",
-                "select_targets_tab": "Categories",
-                "replace_registrants_tab": "Remaining",
-                "replace_targets_tab": "Replacement Categories",
-                "already_selected_tab": "Selected",
-                "id_column": "test_id",
-                "check_same_address": "y",  # Enabled
-                "check_same_address_cols_string": "",  # Empty - should fail validation
-                "columns_to_keep_string": "first_name, last_name",
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet"),
-            },
-        )
-
-        # Should succeed and redirect
-        assert response.status_code == 302
-        assert f"/assemblies/{existing_assembly.id}/data" in response.location
-
-    def test_create_gsheet_hard_validation_passes_when_check_address_disabled(self, logged_in_admin, existing_assembly):
-        """Test hard validation: check_same_address=False allows empty address columns."""
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/validation_pass123456789/edit",
-                "team": "other",  # Custom config
-                "select_registrants_tab": "Respondents",
-                "select_targets_tab": "Categories",
-                "replace_registrants_tab": "Remaining",
-                "replace_targets_tab": "Replacement Categories",
-                "already_selected_tab": "Selected",
-                "id_column": "test_id",
-                # check_same_address not set - defaults to False
-                "check_same_address_cols_string": "",  # Empty - should be OK since check_same_address is False
-                "columns_to_keep_string": "first_name, last_name",
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet"),
-            },
-            follow_redirects=False,
-        )
-
-        # Should succeed and redirect
-        assert response.status_code == 302
-        assert f"/assemblies/{existing_assembly.id}/data" in response.location
-
-    def test_create_gsheet_soft_validation_empty_columns_to_keep_shows_warning(
-        self, logged_in_admin, existing_assembly
-    ):
-        """Test soft validation: empty columns_to_keep shows warning but allows save."""
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/warning123456789/edit",
-                "team": "other",  # Custom config
-                "select_registrants_tab": "Respondents",
-                "select_targets_tab": "Categories",
-                "replace_registrants_tab": "Remaining",
-                "replace_targets_tab": "Replacement Categories",
-                "already_selected_tab": "Selected",
-                "id_column": "test_id",
-                "check_same_address": "y",
-                "check_same_address_cols_string": "address1, postcode",
-                "columns_to_keep_string": "",  # Empty - should show warning
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet"),
-            },
-            follow_redirects=True,  # Follow redirects to check flash messages
-        )
-
-        # Should succeed (redirect was followed, so status is 200)
-        assert response.status_code == 200
-
-        # Check for both success and warning flash messages in the response
-        assert b"configuration created successfully" in response.data or b"success" in response.data.lower()
-        assert (
-            b"Warning" in response.data
-            or b"No columns to keep" in response.data
-            or b"participant data columns" in response.data
-        )
-
-    def test_create_gsheet_soft_validation_empty_columns_to_keep_with_team_does_not_show_warning(
-        self, logged_in_admin, existing_assembly
-    ):
-        """
-        Test soft validation:
-        - empty columns_to_keep would normally show warning but allows save.
-        - team "uk" means the columns_to_keep will be filled in after form validation.
-        """
-        response = logged_in_admin.post(
-            f"/assemblies/{existing_assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/warning123456789/edit",
-                "team": "uk",  # Custom config
-                "select_registrants_tab": "Respondents",
-                "select_targets_tab": "Categories",
-                "replace_registrants_tab": "Remaining",
-                "replace_targets_tab": "Replacement Categories",
-                "already_selected_tab": "Selected",
-                "id_column": "test_id",
-                "check_same_address": "y",
-                "check_same_address_cols_string": "address1, postcode",
-                "columns_to_keep_string": "",  # Empty - should show warning
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{existing_assembly.id}/gsheet"),
-            },
-            follow_redirects=True,  # Follow redirects to check flash messages
-        )
-
-        # Should succeed (redirect was followed, so status is 200)
-        assert response.status_code == 200
-
-        # Check for both success and warning flash messages in the response
-        assert b"configuration created successfully" in response.data or b"success" in response.data.lower()
-        assert (
-            b"Warning" not in response.data
-            and b"No columns to keep" not in response.data
-            and b"participant data columns" not in response.data
-        )
-
-    def test_edit_gsheet_hard_validation_check_address_without_columns(self, logged_in_admin, assembly_with_gsheet):
-        """Test hard validation on edit: check_same_address=True requires address columns."""
-        assembly, _gsheet = assembly_with_gsheet
-
-        response = logged_in_admin.post(
-            f"/assemblies/{assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/edit_validation123456789/edit",
-                "team": "other",  # Custom config
-                "select_registrants_tab": "Respondents",
-                "select_targets_tab": "Categories",
-                "replace_registrants_tab": "Remaining",
-                "replace_targets_tab": "Replacement Categories",
-                "already_selected_tab": "Selected",
-                "id_column": "test_id",
-                "check_same_address": "y",  # Enabled
-                "check_same_address_cols_string": "",  # Empty - should fail validation
-                "columns_to_keep_string": "first_name, last_name",
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{assembly.id}/gsheet"),
-            },
-        )
-
-        # Should return form with validation error (not redirect)
-        assert response.status_code == 200
-        # Check for validation error message
-        assert b"address columns" in response.data.lower() or b"must specify" in response.data.lower()
-
-    def test_edit_gsheet_soft_validation_empty_columns_to_keep_shows_warning(
-        self, logged_in_admin, assembly_with_gsheet
-    ):
-        """Test soft validation on edit: empty columns_to_keep shows warning but allows save."""
-        assembly, _gsheet = assembly_with_gsheet
-
-        response = logged_in_admin.post(
-            f"/assemblies/{assembly.id}/gsheet",
-            data={
-                "url": "https://docs.google.com/spreadsheets/d/edit_warning123456789/edit",
-                "team": "other",  # Custom config
-                "select_registrants_tab": "Respondents",
-                "select_targets_tab": "Categories",
-                "replace_registrants_tab": "Remaining",
-                "replace_targets_tab": "Replacement Categories",
-                "already_selected_tab": "Selected",
-                "id_column": "test_id",
-                "check_same_address": "y",
-                "check_same_address_cols_string": "address1, postcode",
-                "columns_to_keep_string": "",  # Empty - should show warning
-                "csrf_token": get_csrf_token(logged_in_admin, f"/assemblies/{assembly.id}/gsheet"),
-            },
-            follow_redirects=True,  # Follow redirects to check flash messages
-        )
-
-        # Should succeed (redirect was followed, so status is 200)
-        assert response.status_code == 200
-
-        # Check for both success and warning flash messages in the response
-        assert b"configuration updated successfully" in response.data or b"success" in response.data.lower()
-        assert (
-            b"Warning" in response.data
-            or b"No columns to keep" in response.data
-            or b"participant data columns" in response.data
-        )
