+++ backend/tests/e2e/test_backoffice_respondent_field_schema.py
@@ -1,5 +1,7 @@
-"""ABOUTME: End-to-end tests for the respondent field schema management UI
-ABOUTME: Covers view, edit label/group, reorder, delete, and initialise flows"""
+"""ABOUTME: End-to-end PostgreSQL happy-path smokes for the respondent field schema UI
+ABOUTME: Behavioural coverage (validation, render, transitions) lives in tests/component/"""
+
+import pytest
 
 from opendlp.domain.respondent_field_schema import (
     ChoiceOption,
@@ -43,19 +45,6 @@ class TestViewSchemaPage:
         assert b"Fixed" in body
         # Initialise button is not shown when schema already exists.
         assert b"Initialise empty schema" not in body
-
-    def test_renders_initialise_button_when_no_schema(self, logged_in_admin, existing_assembly, admin_user):
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        assert b"Initialise empty schema" in response.data
-
-    def test_redirects_when_not_logged_in(self, client, existing_assembly):
-        response = client.get(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-        assert "login" in response.location
 
 
 class TestInitialiseSchema:
@@ -108,48 +97,8 @@ class TestUpdateField:
             assert moved.group == RespondentFieldGroup.ABOUT_YOU
 
 
-class TestOnRegistrationPage:
-    def test_schema_page_renders_registration_column(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        body = response.get_data(as_text=True)
-        assert "On registration form" in body
-        assert 'name="on_registration_page"' in body
-
-    def test_update_sets_on_registration_page(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            custom_field = next(f for f in schema if f.field_key == "custom_notes")
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{custom_field.id}/update",
-            data={
-                "label": "Custom notes",
-                "on_registration_page": FieldOnRegistrationPage.NO.value,
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            updated = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            field = next(f for f in updated if f.field_key == "custom_notes")
-            assert field.on_registration_page == FieldOnRegistrationPage.NO
-
-
 class TestMoveField:
+    @pytest.mark.db_semantics
     def test_move_up_swaps_with_previous_field(
         self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
     ):
@@ -183,36 +132,6 @@ class TestMoveField:
             after = [f.field_key for f in schema if f.group == RespondentFieldGroup.OTHER]
             assert after == ["b", "a", "c"]
 
-    def test_move_up_at_top_is_noop(self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            import_respondents_from_csv(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                "external_id,a,b\nR001,1,2\n",
-                replace_existing=True,
-            )
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            top = next(f for f in schema if f.field_key == "a")
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{top.id}/move",
-            data={
-                "direction": "up",
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            after = [f.field_key for f in schema if f.group == RespondentFieldGroup.OTHER]
-            assert after == ["a", "b"]
-
 
 class TestAddField:
     def test_add_field_creates_row(self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory):
@@ -245,97 +164,6 @@ class TestAddField:
             assert added.field_type == FieldType.TEXT
             assert added.on_registration_page == FieldOnRegistrationPage.YES_OPTIONAL
 
-    def test_add_choice_field_seeds_placeholder_option(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/add",
-            data={
-                "field_key": "preferred_contact",
-                "field_type": FieldType.CHOICE_RADIO.value,
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            added = next(f for f in schema if f.field_key == "preferred_contact")
-            assert added.field_type == FieldType.CHOICE_RADIO
-            assert [opt.value for opt in added.options] == ["option_1"]
-
-    def test_add_duplicate_key_is_rejected(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            before = len(respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id))
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/add",
-            data={
-                "field_key": "custom_notes",
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=True,
-        )
-        assert response.status_code == 200
-        # The translated conflict message (a LazyString) renders via flash(str(e)).
-        assert b"already exists" in response.data
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            after = len(respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id))
-            assert after == before
-
-    def test_add_empty_key_is_rejected(self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            before = len(respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id))
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/add",
-            data={
-                "field_key": "!!!",
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=True,
-        )
-        assert response.status_code == 200
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            after = len(respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id))
-            assert after == before
-
-    def test_add_form_renders_when_schema_exists(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        assert b"Add a field" in response.data
-        assert f"/assembly/{existing_assembly.id}/respondent-schema/fields/add".encode() in response.data
-
-    def test_add_form_absent_without_schema(self, logged_in_admin, existing_assembly):
-        # No schema yet: the Initialise prompt shows instead of the add form.
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        assert b"Add a field" not in response.data
-
 
 class TestDeleteField:
     def test_delete_non_fixed_field(self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory):
@@ -360,112 +188,8 @@ class TestDeleteField:
             schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
             assert "custom_notes" not in {f.field_key for f in schema}
 
-    def test_cannot_delete_fixed_field(self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            fixed_field = next(f for f in schema if f.is_fixed)
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{fixed_field.id}/delete",
-            data={
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=True,
-        )
-        assert response.status_code == 200
-        # Fixed field is still there.
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            assert fixed_field.field_key in {f.field_key for f in schema}
-
-
-class TestFieldsTab:
-    def test_fields_tab_appears_in_assembly_tab_bar(self, logged_in_admin, existing_assembly):
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/data?source=csv")
-        assert response.status_code == 200
-        assert f"/assembly/{existing_assembly.id}/respondent-schema".encode() in response.data
-
-    def test_fields_tab_is_never_disabled(self, logged_in_admin, existing_assembly):
-        # Visit the Data tab before any data source is chosen — the Fields tab
-        # must still be a live link, not a disabled placeholder.
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/data")
-        assert response.status_code == 200
-        assert f"/assembly/{existing_assembly.id}/respondent-schema".encode() in response.data
-
 
 class TestFieldTypeAndOptions:
-    def test_schema_page_renders_type_column(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        body = response.data
-        assert b"Type" in body
-        assert b"Yes / No / Not set" in body  # fixed flags render as BOOL_OR_NONE
-        assert b"Email" in body  # email fixed row renders as EMAIL type
-
-    def test_update_accepts_field_type(self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            custom = next(f for f in schema if f.field_key == "custom_notes")
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{custom.id}/update",
-            data={
-                "label": custom.label,
-                "group": custom.group.value,
-                "field_type": FieldType.LONGTEXT.value,
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            updated = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            field = next(f for f in updated if f.field_key == "custom_notes")
-            assert field.field_type == FieldType.LONGTEXT
-
-    def test_changing_to_choice_seeds_default_option(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            custom = next(f for f in schema if f.field_key == "custom_notes")
-
-        logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{custom.id}/update",
-            data={
-                "label": custom.label,
-                "group": custom.group.value,
-                "field_type": FieldType.CHOICE_RADIO.value,
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            updated = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            field = next(f for f in updated if f.field_key == "custom_notes")
-            assert field.field_type == FieldType.CHOICE_RADIO
-            assert field.options is not None
-            assert len(field.options) >= 1
-
     def test_add_option_appends_to_choice_field(
         self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
     ):
@@ -503,147 +227,6 @@ class TestFieldTypeAndOptions:
             assert [o.value for o in field.options] == ["initial", "second"]
             assert field.options[1].help_text == "second option"
 
-    def test_remove_option_drops_from_choice_field(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            custom = next(f for f in schema if f.field_key == "custom_notes")
-            respondent_field_schema_service.update_field(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                custom.id,
-                field_type=FieldType.CHOICE_RADIO,
-                options=[ChoiceOption(value="a"), ChoiceOption(value="b")],
-            )
-
-        logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{custom.id}/options/remove",
-            data={
-                "value": "a",
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            updated = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            field = next(f for f in updated if f.field_key == "custom_notes")
-            assert field.options is not None
-            assert [o.value for o in field.options] == ["b"]
-
-    def test_update_option_changes_value_and_help_text(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            custom = next(f for f in schema if f.field_key == "custom_notes")
-            respondent_field_schema_service.update_field(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                custom.id,
-                field_type=FieldType.CHOICE_RADIO,
-                options=[
-                    ChoiceOption(value="old", help_text="original help"),
-                    ChoiceOption(value="other"),
-                ],
-            )
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{custom.id}/options/update",
-            data={
-                "old_value": "old",
-                "value": "renamed",
-                "help_text": "updated help",
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            updated = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            field = next(f for f in updated if f.field_key == "custom_notes")
-            assert field.options is not None
-            assert [o.value for o in field.options] == ["renamed", "other"]
-            assert field.options[0].help_text == "updated help"
-
-    def test_update_option_renders_edit_form(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            custom = next(f for f in schema if f.field_key == "custom_notes")
-            respondent_field_schema_service.update_field(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                custom.id,
-                field_type=FieldType.CHOICE_RADIO,
-                options=[ChoiceOption(value="one", help_text="first option")],
-            )
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        body = response.data
-        # The option value and help_text should both appear as editable input values.
-        assert b'value="one"' in body
-        assert b'value="first option"' in body
-        # The update action URL should be wired up.
-        assert f"/respondent-schema/fields/{custom.id}/options/update".encode() in body
-
-    def test_guess_button_shown_when_conditions_met(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        assert b"Guess field types from data" in response.data
-
-    def test_guess_button_hidden_when_no_respondents(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent_field_schema_service.initialise_empty_schema(uow, admin_user.id, existing_assembly.id)
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        assert b"Guess field types from data" not in response.data
-
-    def test_guess_button_hidden_when_no_text_fields(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            # Flip every non-fixed row away from TEXT so there's nothing to guess.
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            for f in schema:
-                if not f.is_fixed and f.field_type == FieldType.TEXT:
-                    respondent_field_schema_service.update_field(
-                        uow,
-                        admin_user.id,
-                        existing_assembly.id,
-                        f.id,
-                        field_type=FieldType.LONGTEXT,
-                    )
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondent-schema")
-        assert response.status_code == 200
-        assert b"Guess field types from data" not in response.data
-
     def test_guess_post_round_trips_and_flashes(
         self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
     ):
@@ -672,70 +255,3 @@ class TestFieldTypeAndOptions:
             schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
             voted = next(f for f in schema if f.field_key == "voted")
             assert voted.field_type == FieldType.BOOL_OR_NONE
-
-    def test_switch_choice_back_to_text_clears_options(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            custom = next(f for f in schema if f.field_key == "custom_notes")
-            respondent_field_schema_service.update_field(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                custom.id,
-                field_type=FieldType.CHOICE_RADIO,
-                options=[ChoiceOption(value="a"), ChoiceOption(value="b")],
-            )
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{custom.id}/update",
-            data={
-                "label": custom.label,
-                "group": custom.group.value,
-                "field_type": FieldType.TEXT.value,
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            updated = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            field = next(f for f in updated if f.field_key == "custom_notes")
-            assert field.field_type == FieldType.TEXT
-            assert field.options is None
-
-    def test_fixed_row_rejects_field_type_change(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            _seed_schema(uow, admin_user, existing_assembly)
-            schema = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            email_field = next(f for f in schema if f.field_key == "email")
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/respondent-schema/fields/{email_field.id}/update",
-            data={
-                "label": email_field.label,
-                "group": email_field.group.value,
-                "field_type": FieldType.TEXT.value,
-                "csrf_token": get_csrf_token(
-                    logged_in_admin,
-                    f"/backoffice/assembly/{existing_assembly.id}/respondent-schema",
-                ),
-            },
-            follow_redirects=False,
-        )
-        # Redirect back with a flash; type change is silently ignored because is_fixed.
-        assert response.status_code == 302
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            updated = respondent_field_schema_service.get_schema(uow, admin_user.id, existing_assembly.id)
-            email_field = next(f for f in updated if f.field_key == "email")
-            # Still EMAIL — the attempt was refused.
-            assert email_field.field_type == FieldType.EMAIL
