+++ backend/tests/e2e/test_backoffice_respondents.py
@@ -1,7 +1,6 @@
-"""ABOUTME: End-to-end tests for the backoffice respondents blueprint
-ABOUTME: Tests uploading/deleting respondents, viewing the respondents page, and viewing a single respondent"""
+"""ABOUTME: End-to-end smoke tests for the backoffice respondents blueprint
+ABOUTME: One real-DB smoke per route plus the Redis-stashed CSV diff-confirmation flow"""
 
-import uuid
 from datetime import UTC, datetime, timedelta
 from io import BytesIO
 
@@ -13,7 +12,7 @@ from opendlp.domain.respondents import Respondent
 from opendlp.domain.value_objects import RespondentStatus
 from opendlp.service_layer.assembly_service import create_assembly
 from opendlp.service_layer.respondent_field_schema_service import get_schema, initialise_empty_schema
-from opendlp.service_layer.respondent_service import create_respondent, delete_respondent, import_respondents_from_csv
+from opendlp.service_layer.respondent_service import create_respondent
 from opendlp.service_layer.unit_of_work import SqlAlchemyUnitOfWork
 from tests.e2e.helpers import get_csrf_token
 
@@ -31,7 +30,6 @@ def assembly_with_respondents(postgres_session_factory, admin_user):
         )
         assembly_id = assembly.id
 
-    # Add respondents
     with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
         respondents = []
         for i in range(5):
@@ -50,7 +48,7 @@ def assembly_with_respondents(postgres_session_factory, admin_user):
 
 
 class TestBackofficeUploadRespondents:
-    """Test CSV respondent upload in backoffice respondents blueprint."""
+    """CSV respondent upload smoke."""
 
     def test_upload_respondents_with_id_column(
         self,
@@ -89,113 +87,9 @@ class TestBackofficeUploadRespondents:
                 assert "age" in r.attributes
                 assert "person_id" not in r.attributes  # id_column should not be in attributes
 
-    def test_upload_respondents_without_id_column_uses_first_column(
-        self,
-        logged_in_admin,
-        existing_assembly: Assembly,
-        postgres_session_factory,
-    ):
-        """Test uploading respondents CSV without id_column uses first column as ID."""
-        csv_content = "participant_id,name,city\nID001,Charlie,London\nID002,Diana,Paris"
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/data/upload-respondents",
-            data={
-                "file": (BytesIO(csv_content.encode()), "respondents.csv"),
-                "id_column": "",  # Empty means use first column
-                "csrf_token": get_csrf_token(
-                    logged_in_admin, f"/backoffice/assembly/{existing_assembly.id}/data?source=csv"
-                ),
-            },
-            content_type="multipart/form-data",
-            follow_redirects=False,
-        )
-
-        assert response.status_code == 302
-
-        # Verify respondents were created using first column as external_id
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondents = uow.respondents.get_by_assembly_id(existing_assembly.id)
-            assert len(respondents) == 2
-            external_ids = {r.external_id for r in respondents}
-            assert external_ids == {"ID001", "ID002"}
-            # First column should not be in attributes
-            for r in respondents:
-                assert "participant_id" not in r.attributes
-                assert "name" in r.attributes
-                assert "city" in r.attributes
-
-    def test_upload_respondents_with_invalid_id_column_shows_error(
-        self,
-        logged_in_admin,
-        existing_assembly: Assembly,
-    ):
-        """Test uploading respondents CSV with non-existent id_column shows error."""
-        csv_content = "name,email,age\nAlice,alice@example.com,30"
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/data/upload-respondents",
-            data={
-                "file": (BytesIO(csv_content.encode()), "respondents.csv"),
-                "id_column": "nonexistent_column",
-                "csrf_token": get_csrf_token(
-                    logged_in_admin, f"/backoffice/assembly/{existing_assembly.id}/data?source=csv"
-                ),
-            },
-            content_type="multipart/form-data",
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        # Should show error message about invalid column
-        assert b"nonexistent_column" in response.data or b"Invalid CSV" in response.data
-
-    def test_upload_respondents_shows_success_message(
-        self,
-        logged_in_admin,
-        existing_assembly: Assembly,
-    ):
-        """Test that successful upload shows success flash message."""
-        csv_content = "id,name\n1,Test User"
-
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{existing_assembly.id}/data/upload-respondents",
-            data={
-                "file": (BytesIO(csv_content.encode()), "respondents.csv"),
-                "id_column": "",
-                "csrf_token": get_csrf_token(
-                    logged_in_admin, f"/backoffice/assembly/{existing_assembly.id}/data?source=csv"
-                ),
-            },
-            content_type="multipart/form-data",
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        assert b"success" in response.data.lower() or b"uploaded" in response.data.lower()
-
-    def test_upload_respondents_redirects_when_not_logged_in(
-        self,
-        client,
-        existing_assembly: Assembly,
-    ):
-        """Test that unauthenticated users are redirected to login."""
-        csv_content = "id,name\n1,Test"
-
-        response = client.post(
-            f"/backoffice/assembly/{existing_assembly.id}/data/upload-respondents",
-            data={
-                "file": (BytesIO(csv_content.encode()), "respondents.csv"),
-            },
-            content_type="multipart/form-data",
-        )
-
-        assert response.status_code == 302
-        assert "login" in response.location
-
 
 class TestBackofficeDeleteRespondents:
-    """Test CSV respondent deletion in backoffice respondents blueprint."""
+    """CSV respondent bulk deletion smoke."""
 
     def test_delete_respondents_success(
         self,
@@ -227,28 +121,9 @@ class TestBackofficeDeleteRespondents:
             respondents = uow.respondents.get_by_assembly_id(assembly_with_respondents.id)
             assert len(respondents) == 0
 
-    def test_delete_respondents_shows_success_message(
-        self,
-        logged_in_admin,
-        assembly_with_respondents: Assembly,
-    ):
-        """Test that successful respondents deletion shows success flash message."""
-        response = logged_in_admin.post(
-            f"/backoffice/assembly/{assembly_with_respondents.id}/data/delete-respondents",
-            data={
-                "csrf_token": get_csrf_token(
-                    logged_in_admin, f"/backoffice/assembly/{assembly_with_respondents.id}/data?source=csv"
-                ),
-            },
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        assert b"deleted" in response.data.lower() or b"success" in response.data.lower()
-
 
 class TestUploadDiffConfirmation:
-    """Test the schema-diff confirmation flow on CSV re-upload."""
+    """The schema-diff confirmation flow on CSV re-upload (Redis-stashed pending upload)."""
 
     def _upload(self, client, assembly_id, csv_text, filename="respondents.csv"):
         return client.post(
@@ -261,29 +136,6 @@ class TestUploadDiffConfirmation:
             content_type="multipart/form-data",
             follow_redirects=False,
         )
-
-    def test_first_upload_imports_directly_no_diff_page(
-        self, logged_in_admin, existing_assembly, postgres_session_factory
-    ):
-        """An assembly with no schema yet skips the diff page entirely."""
-        response = self._upload(
-            logged_in_admin,
-            existing_assembly.id,
-            "external_id,first_name\nR001,Alice\n",
-        )
-        assert response.status_code == 302
-        # Goes back to data tab, not the confirm-diff page.
-        assert "confirm-diff" not in response.location
-        assert "source=csv" in response.location
-
-    def test_re_upload_with_unchanged_columns_skips_diff(
-        self, logged_in_admin, existing_assembly, postgres_session_factory
-    ):
-        """If the new CSV has identical columns the diff is empty so no confirmation."""
-        self._upload(logged_in_admin, existing_assembly.id, "external_id,first_name\nR001,Alice\n")
-        response = self._upload(logged_in_admin, existing_assembly.id, "external_id,first_name\nR002,Bob\n")
-        assert response.status_code == 302
-        assert "confirm-diff" not in response.location
 
     def test_re_upload_with_added_column_redirects_to_diff_page(
         self, logged_in_admin, existing_assembly, postgres_session_factory
@@ -417,7 +269,7 @@ class TestUploadDiffConfirmation:
 
 
 class TestBackofficeViewRespondentsPage:
-    """Test backoffice respondents list page."""
+    """Backoffice respondents list page smoke."""
 
     def test_view_respondents_page_with_csv_source(
         self,
@@ -431,100 +283,9 @@ class TestBackofficeViewRespondentsPage:
         # Should see the page content
         assert b"Respondents" in response.data or b"respondents" in response.data.lower()
 
-    def test_view_respondents_page_redirects_when_not_logged_in(
-        self,
-        client,
-        assembly_with_respondents: Assembly,
-    ):
-        """Test that unauthenticated users are redirected to login."""
-        response = client.get(f"/backoffice/assembly/{assembly_with_respondents.id}/respondents")
-
-        assert response.status_code == 302
-        assert "login" in response.location
-
-    def test_view_respondents_page_lists_each_respondent(
-        self,
-        logged_in_admin,
-        existing_assembly: Assembly,
-        admin_user,
-        postgres_session_factory,
-    ):
-        """Table should show external_id, email, and display name for every respondent."""
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondents = [
-                Respondent(
-                    assembly_id=existing_assembly.id,
-                    external_id=f"R-{i:03d}",
-                    email=f"person{i}@example.com",
-                    attributes={"first_name": f"First{i}", "last_name": f"Last{i}"},
-                )
-                for i in range(3)
-            ]
-            uow.respondents.bulk_add(respondents)
-            uow.commit()
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents")
-
-        assert response.status_code == 200
-        body = response.data
-        for i in range(3):
-            assert f"R-{i:03d}".encode() in body
-            assert f"person{i}@example.com".encode() in body
-            assert f"First{i} Last{i}".encode() in body
-
-    def test_view_respondents_page_has_view_link_per_respondent(
-        self,
-        logged_in_admin,
-        existing_assembly: Assembly,
-        postgres_session_factory,
-    ):
-        """Each row should link to the single-respondent page."""
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondents = [
-                Respondent(
-                    assembly_id=existing_assembly.id,
-                    external_id=f"R-{i}",
-                    email=f"p{i}@example.com",
-                    attributes={"name": f"Person {i}"},
-                )
-                for i in range(2)
-            ]
-            uow.respondents.bulk_add(respondents)
-            uow.commit()
-            respondent_ids = [r.id for r in respondents]
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents")
-
-        assert response.status_code == 200
-        for respondent_id in respondent_ids:
-            expected = f"/backoffice/assembly/{existing_assembly.id}/respondents/{respondent_id}"
-            assert expected.encode() in response.data
-
-    def test_view_respondents_page_falls_back_to_email_local_part_when_no_name_fields(
-        self,
-        logged_in_admin,
-        existing_assembly: Assembly,
-        postgres_session_factory,
-    ):
-        """Display name column should show email local-part when no recognisable name attributes exist."""
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = Respondent(
-                assembly_id=existing_assembly.id,
-                external_id="R-EMAIL",
-                email="jane.doe@example.com",
-                attributes={"age": "42"},
-            )
-            uow.respondents.bulk_add([respondent])
-            uow.commit()
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents")
-
-        assert response.status_code == 200
-        assert b"jane.doe" in response.data
-
 
 class TestBackofficeViewSingleRespondent:
-    """Test backoffice view-single-respondent page."""
+    """Backoffice view-single-respondent page smoke."""
 
     def test_view_respondent_renders_for_admin(
         self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
@@ -545,187 +306,14 @@ class TestBackofficeViewSingleRespondent:
         assert b"alice@example.com" in response.data
         assert b"Pool" in response.data
 
-    def test_view_respondent_requires_login(self, client, existing_assembly, admin_user, postgres_session_factory):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R001",
-                attributes={},
-                email="alice@example.com",
-            )
-
-        response = client.get(f"/backoffice/assembly/{existing_assembly.id}/respondents/{respondent.id}")
-        assert response.status_code == 302
-        assert "login" in response.location
-
-    def test_view_respondent_not_found(self, logged_in_admin, existing_assembly):
-        bogus_id = uuid.uuid4()
-        response = logged_in_admin.get(
-            f"/backoffice/assembly/{existing_assembly.id}/respondents/{bogus_id}",
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-        with logged_in_admin.session_transaction() as session:
-            flash_messages = [msg[1] for msg in session.get("_flashes", [])]
-            assert any("Respondent not found" in msg for msg in flash_messages)
-
-    def test_view_respondent_shows_generated_name_from_attributes(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R001",
-                attributes={"first_name": "Sarah", "last_name": "Jones"},
-                email="sarah@example.com",
-            )
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents/{respondent.id}")
-        assert response.status_code == 200
-        # Generated name appears in h1 and title
-        assert b"Sarah Jones" in response.data
-        # External id still present in the details table
-        assert b"R001" in response.data
-
-    def test_view_respondent_falls_back_to_email_local_part(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R001",
-                attributes={},
-                email="sarah.jones@example.com",
-            )
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents/{respondent.id}")
-        assert response.status_code == 200
-        # No name attributes → h1 uses email local-part
-        assert b"sarah.jones" in response.data
-
-    def test_view_respondent_renders_grouped_sections_after_csv_import(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        csv_content = (
-            "external_id,first_name,last_name,gender,postcode,custom_notes\n"
-            "R001,Alice,Jones,Female,SW1A 1AA,extra-info\n"
-        )
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            import_respondents_from_csv(uow, admin_user.id, existing_assembly.id, csv_content)
-            respondent = uow.respondents.get_by_external_id(existing_assembly.id, "R001")
-            assert respondent is not None
-            respondent_id = respondent.id
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents/{respondent_id}")
-        assert response.status_code == 200
-        body = response.data
-        # Status block: external_id and selection status tag.
-        assert b"R001" in body
-        assert b"govuk-tag" in body
-        # Schema-driven group headings are present.
-        assert b"Name and contact" in body
-        assert b"About you" in body
-        assert b"Address" in body
-        # Non-fixed values render via attributes lookup.
-        assert b"Alice" in body
-        assert b"Jones" in body
-        assert b"Female" in body
-        assert b"SW1A 1AA" in body
-        assert b"extra-info" in body
-        # Activity section replaces the old Record metadata block.
-        assert b"Activity" in body
-        assert b"<details" in body
-
-    def test_view_respondent_wrong_assembly(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            other_assembly = create_assembly(
-                uow=uow,
-                title="Other Assembly",
-                created_by_user_id=admin_user.id,
-                question="Other?",
-                first_assembly_date=(datetime.now(UTC).date() + timedelta(days=30)),
-            )
-            other_id = other_assembly.id
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                other_id,
-                external_id="R-OTHER",
-                attributes={},
-            )
-
-        response = logged_in_admin.get(
-            f"/backoffice/assembly/{existing_assembly.id}/respondents/{respondent.id}",
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-        with logged_in_admin.session_transaction() as session:
-            flash_messages = [msg[1] for msg in session.get("_flashes", [])]
-            assert any("Respondent not found" in msg for msg in flash_messages)
-
-
-class TestViewRespondentDeletionUI:
-    """The delete form and DELETED banner on the single-respondent page."""
-
-    def test_admin_sees_delete_form(self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-DEL",
-                attributes={},
-                email="x@example.com",
-            )
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents/{respondent.id}")
-        assert response.status_code == 200
-        assert b"Delete personal data" in response.data
-
-    def test_deleted_respondent_shows_banner_and_comment_list(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-DEAD",
-                attributes={"Gender": "Female"},
-                email="x@example.com",
-            )
-
-        # Delete it via the service so the comment is real
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            delete_respondent(uow, admin_user.id, existing_assembly.id, respondent.id, comment="gdpr")
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents/{respondent.id}")
-        assert response.status_code == 200
-        # Banner
-        assert b"Personal data deleted" in response.data
-        # Comment list row with text
-        assert b"gdpr" in response.data
-        # Form should not render for DELETED respondents
-        assert b"Confirm delete" not in response.data
-
 
 class TestDeleteRespondentRoute:
-    """POST /backoffice/assembly/<id>/respondents/<id>/delete"""
+    """POST /backoffice/assembly/<id>/respondents/<id>/delete smoke."""
 
-    def _delete_url(self, assembly_id: uuid.UUID, respondent_id: uuid.UUID) -> str:
+    def _delete_url(self, assembly_id, respondent_id) -> str:
         return f"/backoffice/assembly/{assembly_id}/respondents/{respondent_id}/delete"
 
-    def _view_url(self, assembly_id: uuid.UUID, respondent_id: uuid.UUID) -> str:
+    def _view_url(self, assembly_id, respondent_id) -> str:
         return f"/backoffice/assembly/{assembly_id}/respondents/{respondent_id}"
 
     def test_admin_can_delete_with_comment(
@@ -761,95 +349,9 @@ class TestDeleteRespondentRoute:
             delete_comments = [c for c in reloaded.comments if c.text == "gdpr request"]
             assert len(delete_comments) == 1
 
-    def test_missing_comment_rejected(self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-KEEP",
-                attributes={},
-                email="keep@example.com",
-            )
-
-        response = logged_in_admin.post(
-            self._delete_url(existing_assembly.id, respondent.id),
-            data={
-                "comment": "",
-                "csrf_token": get_csrf_token(logged_in_admin, self._view_url(existing_assembly.id, respondent.id)),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            reloaded = uow.respondents.get(respondent.id)
-            assert reloaded is not None
-            assert reloaded.selection_status.value == "POOL"
-            assert reloaded.email == "keep@example.com"
-            # Only the CREATE comment from creation; no DELETE was applied.
-            assert all(c.action.value != "DELETE" for c in reloaded.comments)
-
-        with logged_in_admin.session_transaction() as session:
-            flash_messages = [msg[1] for msg in session.get("_flashes", [])]
-            assert any("comment is required" in msg.lower() for msg in flash_messages)
-
-    def test_requires_login(self, client, existing_assembly, admin_user, postgres_session_factory):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-NOLOGIN",
-                attributes={},
-            )
-
-        response = client.post(
-            self._delete_url(existing_assembly.id, respondent.id),
-            data={"comment": "gdpr"},
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-        assert "login" in response.location
-
-    def test_respondent_in_other_assembly(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            other_assembly = create_assembly(
-                uow=uow,
-                title="Other Assembly",
-                created_by_user_id=admin_user.id,
-                question="Other?",
-                first_assembly_date=(datetime.now(UTC).date() + timedelta(days=30)),
-            )
-            other_id = other_assembly.id
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                other_id,
-                external_id="R-OTHER",
-                attributes={},
-            )
-
-        response = logged_in_admin.post(
-            self._delete_url(existing_assembly.id, respondent.id),
-            data={
-                "comment": "gdpr",
-                "csrf_token": get_csrf_token(logged_in_admin, f"/backoffice/assembly/{existing_assembly.id}"),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-        with logged_in_admin.session_transaction() as session:
-            flash_messages = [msg[1] for msg in session.get("_flashes", [])]
-            assert any("Respondent not found" in msg for msg in flash_messages)
-
 
 class TestSelectionStatusTransition:
-    """Transition buttons + POST /.../transition-status."""
+    """POST /.../transition-status smoke."""
 
     def _transition_url(self, assembly_id, respondent_id):
         return f"/backoffice/assembly/{assembly_id}/respondents/{respondent_id}/transition-status"
@@ -857,47 +359,9 @@ class TestSelectionStatusTransition:
     def _view_url(self, assembly_id, respondent_id):
         return f"/backoffice/assembly/{assembly_id}/respondents/{respondent_id}"
 
-    def test_view_page_shows_buttons_for_each_allowed_transition(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            resp = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-T1",
-                attributes={},
-                selection_status=RespondentStatus.SELECTED,
-            )
-        response = logged_in_admin.get(self._view_url(existing_assembly.id, resp.id))
-        assert response.status_code == 200
-        assert b"CONFIRMED" in response.data
-        assert b"WITHDRAWN" in response.data
-
-    def test_view_page_hides_buttons_on_terminal_status(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        # DELETED is the only terminal status; all other statuses have manual
-        # transitions back into the active set.
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            resp = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-T2",
-                attributes={},
-                selection_status=RespondentStatus.DELETED,
-            )
-        response = logged_in_admin.get(self._view_url(existing_assembly.id, resp.id))
-        assert response.status_code == 200
-        # No "Change to" buttons (no allowed transitions)
-        assert b"Change to" not in response.data
-
     def test_post_valid_transition_updates_and_flashes(
         self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
     ):
-
         with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
             resp = create_respondent(
                 uow,
@@ -922,72 +386,12 @@ class TestSelectionStatusTransition:
             retrieved = uow.respondents.get(resp.id)
             assert retrieved.selection_status == RespondentStatus.CONFIRMED
 
-    def test_post_blank_comment_rejected(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            resp = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-T4",
-                attributes={},
-                selection_status=RespondentStatus.SELECTED,
-            )
-
-        response = logged_in_admin.post(
-            self._transition_url(existing_assembly.id, resp.id),
-            data={
-                "new_status": RespondentStatus.CONFIRMED.value,
-                "comment": "",
-                "csrf_token": get_csrf_token(logged_in_admin, self._view_url(existing_assembly.id, resp.id)),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            retrieved = uow.respondents.get(resp.id)
-            assert retrieved.selection_status == RespondentStatus.SELECTED
-
-    def test_post_illegal_transition_rejected(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        # Moves into DELETED are still forbidden via this form (GDPR delete
-        # has its own action), so use that as the illegal case.
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            resp = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-T5",
-                attributes={},
-                selection_status=RespondentStatus.SELECTED,
-            )
-
-        response = logged_in_admin.post(
-            self._transition_url(existing_assembly.id, resp.id),
-            data={
-                "new_status": RespondentStatus.DELETED.value,
-                "comment": "try",
-                "csrf_token": get_csrf_token(logged_in_admin, self._view_url(existing_assembly.id, resp.id)),
-            },
-            follow_redirects=False,
-        )
-        assert response.status_code == 302
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            retrieved = uow.respondents.get(resp.id)
-            assert retrieved.selection_status == RespondentStatus.SELECTED
-
 
 class TestEditRespondentPage:
-    """GET /backoffice/assembly/<id>/respondents/<id>/edit and POST handling."""
+    """POST /backoffice/assembly/<id>/respondents/<id>/edit smoke."""
 
-    def _edit_url(self, assembly_id: uuid.UUID, respondent_id: uuid.UUID) -> str:
+    def _edit_url(self, assembly_id, respondent_id) -> str:
         return f"/backoffice/assembly/{assembly_id}/respondents/{respondent_id}/edit"
-
-    def _view_url(self, assembly_id: uuid.UUID, respondent_id: uuid.UUID) -> str:
-        return f"/backoffice/assembly/{assembly_id}/respondents/{respondent_id}"
 
     def _make_respondent(self, uow, admin_user_id, assembly_id, **kwargs):
         initialise_empty_schema(uow, admin_user_id, assembly_id)
@@ -1000,53 +404,6 @@ class TestEditRespondentPage:
             email=kwargs.pop("email", "r@example.com"),
             **kwargs,
         )
-
-    def test_get_renders_form_grouped_by_schema(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            import_respondents_from_csv(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                "external_id,first_name,note\nR1,Alice,orig\n",
-                replace_existing=True,
-            )
-            resp = next(iter(uow.respondents.get_by_assembly_id(existing_assembly.id)))
-            resp_id = resp.id
-
-        response = logged_in_admin.get(self._edit_url(existing_assembly.id, resp_id))
-        assert response.status_code == 200
-        assert b"Edit respondent" in response.data
-        assert b"Change note" in response.data
-        assert b"first_name" in response.data or b"First name" in response.data
-        # Status transition form must also be rendered on the Edit page.
-        assert b"Selection status" in response.data
-        assert b"Change to" in response.data
-
-    def test_get_with_uninitialised_schema_shows_init_prompt(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        """When no field schema has been initialised the form is replaced
-        with a prompt linking to the Fields tab."""
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = create_respondent(
-                uow,
-                admin_user.id,
-                existing_assembly.id,
-                external_id="R-NOSCHEMA",
-                attributes={},
-            )
-            resp_id = respondent.id
-
-        response = logged_in_admin.get(self._edit_url(existing_assembly.id, resp_id))
-        assert response.status_code == 200
-        assert b"This assembly has no respondent fields configured yet." in response.data
-        # Link points at the Fields tab.
-        assert f"/assembly/{existing_assembly.id}/respondent-schema".encode() in response.data
-        # The form itself is not rendered.
-        assert b"Save changes" not in response.data
-        assert b"Change note" not in response.data
 
     def test_post_valid_updates_redirects_and_flashes(
         self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
@@ -1070,59 +427,3 @@ class TestEditRespondentPage:
             assert retrieved is not None
             assert retrieved.email == "new@example.com"
             assert any(c.text == "corrected email" for c in retrieved.comments)
-
-    def test_post_blank_comment_rerenders_with_error(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = self._make_respondent(uow, admin_user.id, existing_assembly.id)
-            resp_id = respondent.id
-
-        response = logged_in_admin.post(
-            self._edit_url(existing_assembly.id, resp_id),
-            data={
-                "email": "other@example.com",
-                "comment": "",
-                "csrf_token": get_csrf_token(logged_in_admin, self._edit_url(existing_assembly.id, resp_id)),
-            },
-            follow_redirects=False,
-        )
-        # Re-renders form (200), email unchanged.
-        assert response.status_code == 200
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            retrieved = uow.respondents.get(resp_id)
-            assert retrieved is not None
-            assert retrieved.email == "r@example.com"
-
-    def test_refused_for_deleted_respondent(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = self._make_respondent(uow, admin_user.id, existing_assembly.id)
-            resp_id = respondent.id
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            delete_respondent(uow, admin_user.id, existing_assembly.id, resp_id, comment="gdpr")
-
-        response = logged_in_admin.get(self._edit_url(existing_assembly.id, resp_id), follow_redirects=False)
-        assert response.status_code == 302
-        assert self._view_url(existing_assembly.id, resp_id) in response.location
-
-    def test_view_page_shows_edit_button(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            respondent = self._make_respondent(uow, admin_user.id, existing_assembly.id)
-            resp_id = respondent.id
-        response = logged_in_admin.get(self._view_url(existing_assembly.id, resp_id))
-        assert response.status_code == 200
-        assert b"/edit" in response.data
-
-    def test_list_page_shows_edit_link_for_each_row(
-        self, logged_in_admin, existing_assembly, admin_user, postgres_session_factory
-    ):
-        with SqlAlchemyUnitOfWork(postgres_session_factory) as uow:
-            self._make_respondent(uow, admin_user.id, existing_assembly.id)
-
-        response = logged_in_admin.get(f"/backoffice/assembly/{existing_assembly.id}/respondents")
-        assert response.status_code == 200
-        assert b"Edit" in response.data
